impl built-in Offsetof for instance of generic function
